# server/openapi.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: realm server
servers:
  - url: http://localhost:8001/swagger
paths:
  /users:
    get:
      tags:
        - User
      operationId: getUserList
      summary: ユーザーリストを取得する
      responses:
        '200':
          description: ユーザーリストを取得しました
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
        '500':
          description: "サーバーエラー"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - User
      operationId: createUser
      summary: ユーザーを作成する
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
        required: true
      responses:
        '201':
          description: ユーザーを作成しました
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        '400':
          description: "バリデーションエラー"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: "サーバーエラー"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    UserInput:
      type: object
      properties:
        name:
          type: string
          x-oapi-codegen-extra-tags:
            validate: required
      required:
        - name
    UserResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - createdAt
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message